CONJ-27 : Implement tcpAbortiveClose option, for "hard" socket closing.
If set, client will abort connection in a way that TCP connection does not enter usual TCP_WAIT state upon close, and free local port. This option can be handy in situation where connections are created and destroyed in rapid succession and new connection can't be created because of port shortage.