Added functions for checking for super privileges and local connections
Added function hasSuperPrivileges to check if the user has super
privileges.

Added function isLocalConnection for checking if the current connection
is to localhost or not. This is useful for e.g. socket and named pipes
tests that only can be run if the connection is local.

Some corrections in test cases where database name was hardcoded and
addition of junit Assume rules.