Add allowLocalInfile connection property defaulting to true.
Adds a new connection property, allowLocalInfile (defaulting to false).
It indicates whether or not the connection should respond to LOCALINFILE
packets when an explicit InputStream for the connection has not been set.
Previously the driver would open whatever local file or URL stream the
server requested it to open. With allowLocalInfile=false the driver will
throw a QueryException if localInfileInputStream is null (ie. it will not
allow reading from the local file system or URL).