AbstractMastersListener: Fix for infinite loop when unscheduling fail loop
The return statement was missing after the unschedule completed, thus resulting in an infinite loop.
Because the task should NOT be terminated except through "stopFailover" the failLoop atomicReference was also made private.  This is critical because calls to "blockTillTerminated" without going through "stopFailover" will also result in an infinite loop.