MastersSlavesListener: Changes how fail loops are controlled and dynamic scheduler is maintained
This fixes a couple race conditions in MastersSlavesListener.
* Under some possible conditions actions could be attempted on a null dynamic sized scheduler.
* It was possible to start too many FailoverLoop runnables.  Because we are checking the size against an already incremented pool size, and because there is no locking between that size and add check.  Several threads may read a size one less than the total desired, but all threads may start those fail loops and then the size will be greater than the maximum count.

The solution to these issues was:
* The pool is no longer ever unset.  It is just a final static instance, we may scale it to zero (and if we do, it will then be elegible for garbage collection).
* FailoverLoop's are increased and decreased from within a single recurring task that runs every 250 mills.  This means that there may be a 250 millisecond delay for failover loops being incremented after a listener is created.

I also changed the DynamicSizedSchedulerInterface to be more generic.  In the future schedulers may be wanted to be used in other places.  It is only MastersSlavesListener that is interested in scaling the thread counts based off listener counts.  So instead that math is done in MastersSlavesListener, and the dynamic scheduler just gets simple pool size requests.

The SchedulerServiceProviderHolderTest for shutdown behavior was removed.  This unit test is no longer valid since it is expected that pools are now shutdown by the normal means.