Rethrow connection exceptions after MariaDbServerPreparedStatement fails.
The fix for CONJ-238 was to catch all SQLExceptions thrown by the constructor
for MariaDbServerPreparedStatement, and fallback to a client prepared
statement. This is fine except for SQLNonTransientConnectionExceptions, which
can be the result of an IOException that leaves PacketOutputStream in an
invalid state.

If that happens, then executing the returned MariaDbClientPreparedStatement
will fail with a mysterious "Last stream not finished" (or, as of 1.3.5,
"Stream has already closed") error, with no indication of what previously went
wrong.