MariaDbStatement: Removed "TimerTask"
Instead of using timers which create threads on demand, this leverages the existing "DynamicSizedScheduler" concept to allow a pool to reuse possible threads.
Thus avoiding the need to create possibly lots of threads if there is a server side issue requiring lots of queries to timeout.  Using this dynamic pool we can potentially reusue threads.