getTableTypes() returns "TABLE" consistent with getTables()
In CONJ-218 (change 4e8c7153e6a7f25128753f2b6527846013e24136), a table type of "BASE TABLE" is translated to "TABLE". The JDBC documentation describes "TABLE" as typical, so that's reasonable.
getTableTypes() returns the set of possible values for table types to be passed as a parameter to getTables() and to be expected in the results from getTables(). It should be modified to match, i.e. it should produce a value of "TABLE" and not "BASE TABLE".