[CONJ-828] new option `ensureSocketState`
When a query is issued, the connector ensures that any streaming result-set will be fully read from the socket before issuing a new command.
This permits to ensure the socket state.

Adding a new option `ensureSocketState` that when enable will ensure that the socket buffer is empty before issuing new command. (this doesn't concern pipelining commands). If data is present in socket. An error will be raised, throwing the content of socket data to permit identification of error. The goal of this option is mainly for debugging.

This functionality add a few overheads so will be disabled by default