[CONJ-810] normalization of resultset getDate/getTime of timestamp field.
To conform with the definition of SQL DATE, the millisecond values wrapped by a java.sql.Date instance is now 'normalized' by setting the hours, minutes, seconds, and milliseconds to zero in the particular time zone with which the instance is associated (current client timezone or if option `useLegacyDatetimeCode` is disabled to server timezone).

Same for java.sql.Time: The date components is set to the "zero epoch" value of January 1, 1970.

example: getTime() on value '2015-03-29T01:45:00.012' return a java.sql.Time  with value '01:45:00', submethod getTime return time in millisecond that did correspond to '2015-03-29T01:45:00.012'. Now correspond to '1970-01-01T01:45:00.012'