Use 32760 as the "unused" port number
Port 3000 is not that uncommon of a port for a MariaDB instance to run on. To avoid conflicting with setups where there actually is a MariaDB instance running on that port, use a port that is near the end of the ephemeral range. Hopefully this is something that is less likely to conflict.